Cleaned up files and added retrieve_dbData.php
Took all of the html and php files that were relevant and lumped them together in a folder. 
Also added the retrieve_dbData.php file which connects to the databse and pulls the algorithm data. Right now it's connected to a dummy database with dummy data since the real one is down right now.